FedRain/pyproject.toml
[tool.poetry]
authors = ["Weiyüen Wu <doomsplayer@gmail.com>"]
description = "Experiments for Rain"
name = "RainExperiment"
version = "0.1.0"

[tool.poetry.dependencies]
matplotlib = "^3"
mnist = "^0.2"
pandas = "^1"
psycopg2-binary = "^2"
python = "^3.7"
scipy = "^1"
sqlalchemy = "^1"
tensorflow-probability = "^0.9"
tqdm = "^4"
# scikit-image = "^0.15"
altair = "^4.0.1"
contexttimer = "^0.3.3"
docopt = "^0.6"
gmpy2 = "^2.0.8"
grpcio-tools = "^1.37.1"
numpy = "^1"
phe = "^1.4.0"
scikit-learn = "^0.22.2"
svglib = "^0.9.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
jupyter = "^1.0.0"
mypy = "^0.770"
pylint = "^2.4.4"

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
